@@ -13,13 +13,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.0.0",
-    "@angular/core": "~13.0.0",
+    "@angular/common": "^14.2.12",
+    "@angular/core": "^14.2.12",
     "@angular/fire": "^7.3.0",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@capacitor/android": "3.5.1",
     "@capacitor/app": "1.1.0",
     "@capacitor/core": "3.5.1",
@@ -52,15 +52,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.1",
+    "@angular-devkit/build-angular": "^14.2.10",
     "@angular-eslint/builder": "~13.0.1",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.0.1",
-    "@angular/compiler": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0",
-    "@angular/language-service": "~13.0.0",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler": "^14.2.12",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@capacitor/cli": "3.5.1",
     "@ionic/angular-toolkit": "^5.0.0",
     "@ngrx/schematics": "^13.0.2",
@@ -83,7 +83,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   },
   "description": "Chess platform"
 }
